Django/requirements.txt
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
gunicorn==21.2.0
psycopg2-binary==2.9.10
cryptography==46.0.3
pika==1.3.2
